common: double width of sUnassociatedStationStats.time_stamp to delay overflow
EasyMesh Unassociated STA Link Metrics Response TLV contains
a 4byte / 32bit field that should hold the time delta
between the 'time point when the measurement was made' and
the 'time point when the Report was sent'

PWHM exposes time as a RFC3339 string
using lib amxc to parse this string into a struct containing
a signed 64bit 'seconds' field.

https://prplfoundationcloud.atlassian.net/browse/PPM-3413

(cherry picked from commit 752fcf491e00ce43a921eab6933b76ec8c21fe19)

+++ common/beerocks/bwl/whm/mon_wlan_hal_whm.cpp
@@ -957,7 +957,7 @@ bool mon_wlan_hal_whm::sta_unassoc_rssi_measurement(
         amxc_ts_t time;
         memset(&time, 0, sizeof(amxc_ts_t));
         amxc_ts_parse(&time, time_stamp_str.c_str(), time_stamp_str.size());
-        uint32_t timestamp_ms = time.sec * 1000 + time.nsec / 1000000;
+        uint64_t timestamp_ms = time.sec * 1000 + time.nsec / 1000000;
         // TimeStamp is datetime as per RFC3339 and maybe contains fractions of seconds information
 
         if (new_list.find(mac_address_amx) != new_list.end()) {
